Fix: use AtomicU32 instead of AtomicU64 for 32-bit target support

+++ agent/src/monitoring/system_impls/mod.rs
@@ -11,7 +11,7 @@ use ng_monitoring::data_structure::{
 };
 use process::count_processes;
 use std::sync::Arc;
-use std::sync::atomic::{AtomicU64, Ordering};
+use std::sync::atomic::{AtomicU32, Ordering};
 use sysinfo::System;
 use tokio::sync::{Mutex, MutexGuard, OnceCell};
 use virtualization_detect::detect_virtualization;
@@ -24,7 +24,12 @@ use virtualization_detect::detect_virtualization;
 const PROCESS_REFRESH_INTERVAL_SECS: u64 = 5;
 
 /// 全局缓存的进程数，由独立后台 ticker 周期更新，dynamic tick 直接读取（无采集开销）。
-static PROCESS_COUNT_CACHE: AtomicU64 = AtomicU64::new(0);
+///
+/// 用 `AtomicU32` 而非 `AtomicU64`：`AtomicU64` 在部分 32-bit target
+///（armv5te / mipsel / powerpc / thumbv6m 等）上不可用，而 `AtomicU32` 在所有
+/// Rust target 上都有保证。`count_processes()` 返回 `u32`，且 `u32::MAX`（≈42.9 亿）
+/// 远超任何真实进程数，故无精度损失。
+static PROCESS_COUNT_CACHE: AtomicU32 = AtomicU32::new(0);
 
 /// 保证进程数采集 ticker 只启动一次。
 static PROCESS_TICKER_STARTED: std::sync::OnceLock<()> = std::sync::OnceLock::new();
@@ -56,7 +61,7 @@ async fn refresh_process_count() {
     let count = tokio::task::spawn_blocking(count_processes)
         .await
         .unwrap_or(0);
-    PROCESS_COUNT_CACHE.store(u64::from(count), Ordering::Relaxed);
+    PROCESS_COUNT_CACHE.store(count, Ordering::Relaxed);
 }
 
 /// 读取缓存的进程数（最近一次采集值，无 `/proc` 遍历开销）。
@@ -64,7 +69,7 @@ async fn refresh_process_count() {
 /// 注：dynamic summary 上报的 `process_count` 是最近一次 ticker 刷新值
 /// （最多滞后 `PROCESS_REFRESH_INTERVAL_SECS` 秒），而非精确到当前秒。
 fn cached_process_count() -> u64 {
-    PROCESS_COUNT_CACHE.load(Ordering::Relaxed)
+    u64::from(PROCESS_COUNT_CACHE.load(Ordering::Relaxed))
 }
 
 /// 获取精确的 OS 版本号。
